third_party/blink/web_tests/http/tests/preload/warning/blob-request.html

<!doctype html>
<html>
<link rel="preload" as="fetch" crossorigin="anonymous" href="resources/hello.txt"></link>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
promise_test((t) => {
  const xhr = new XMLHttpRequest;
  xhr.open('GET', 'resources/hello.txt');
  xhr.responseType = 'blob';
  const promise = new Promise(resolve => {
    xhr.onloadend = resolve;
  });
  xhr.send();
  return promise;
}, 'Preload match fails because the request wants to load the content as a blob (see the console log).');
</script>
</body>
</html>
